Add large shift around and usage of hooks for ItemDetailsModal
